HHH-19370 move SqlAstTranslators to a dedicated subpackage
