+++ hibernate-community-dialects/src/main/java/org/hibernate/community/dialect/TiDBSqlAstTranslator.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 import org.hibernate.LockOptions;
 import org.hibernate.dialect.DmlTargetColumnQualifierSupport;
-import org.hibernate.dialect.MySQLSqlAstTranslator;
+import org.hibernate.dialect.sql.ast.MySQLSqlAstTranslator;
 import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.internal.util.collections.Stack;
 import org.hibernate.query.sqm.ComparisonOperator;
