+++ hibernate-core/src/main/java/org/hibernate/dialect/MariaDBDialect.java
@@ -19,6 +19,7 @@ import org.hibernate.dialect.identity.IdentityColumnSupport;
 import org.hibernate.dialect.identity.MariaDBIdentityColumnSupport;
 import org.hibernate.dialect.sequence.MariaDBSequenceSupport;
 import org.hibernate.dialect.sequence.SequenceSupport;
+import org.hibernate.dialect.sql.ast.MariaDBSqlAstTranslator;
 import org.hibernate.dialect.type.MariaDBCastingJsonArrayJdbcTypeConstructor;
 import org.hibernate.dialect.type.MariaDBCastingJsonJdbcType;
 import org.hibernate.engine.jdbc.dialect.spi.DialectResolutionInfo;
