+++ hibernate-core/src/main/java/org/hibernate/dialect/sql/ast/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Internal package containing dialect-specific extensions of
+ * {@link org.hibernate.sql.ast.spi.AbstractSqlAstTranslator}.
+ */
+@Internal
+package org.hibernate.dialect.sql.ast;
+
+import org.hibernate.Internal;
