+++ hibernate-core/src/main/java/org/hibernate/dialect/type/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Internal package containing dialect-specific implementations of
+ * {@link org.hibernate.annotations.JdbcType}.
+ */
+@Internal
+package org.hibernate.dialect.type;
+
+import org.hibernate.Internal;
