+++ hibernate-core/src/main/java/org/hibernate/sql/ast/spi/package-info.java
@@ -6,4 +6,7 @@
 /**
  * Package defining support for creating and consuming a SQL AST.
  */
+@Incubating
 package org.hibernate.sql.ast.spi;
+
+import org.hibernate.Incubating;
