+++ hibernate-core/src/main/java/org/hibernate/sql/ast/tree/package-info.java
@@ -7,5 +7,10 @@
  * Package defining the SQL AST.
  * <p>
  * The package {@link org.hibernate.sql.ast.spi} defines support for creating and consuming the AST.
+ * <p>
+ * This package and all subpackages are currently incubating and subject to change.
  */
+@Incubating
 package org.hibernate.sql.ast.tree;
+
+import org.hibernate.Incubating;
